refactor: Joined git and serve method
