+++ src/server/api/state.rs
@@ -13,6 +13,8 @@ pub trait Global {
     fn archive(&self) -> &Archive;
     /// Database connection
     fn db(&self) -> &db::DatabaseConnection;
+    /// path to the Stelae archive
+    fn archive_path(&self) -> &PathBuf;
 }
 
 /// Application state
@@ -22,6 +24,8 @@ pub struct App {
     pub archive: Archive,
     /// Database connection
     pub db: db::DatabaseConnection,
+    /// path to the Stelae archive
+    pub archive_path: PathBuf,
 }
 
 impl Global for App {
@@ -31,6 +35,10 @@ impl Global for App {
 
     fn db(&self) -> &db::DatabaseConnection {
         &self.db
+    }
+
+    fn archive_path(&self) -> &PathBuf {
+        &self.archive_path
     }
 }
